CLOUDP-165150: minimize logging in transformation engine
