+++ tools/transformer/src/transform.js
@@ -1,29 +1,21 @@
-const { readFileSync, writeFileSync } = require("fs");
 const {
   applyAllOfTransformations,
   applyOneOfTransformations,
   applyModelNameTransformations,
 } = require("./transformations");
-const { parse, stringify } = require("yaml");
+const { getAPI, saveAPI } = require("./apifile");
 
-let apiFileLocation = process.argv.slice(2);
-console.error(apiFileLocation);
-if (!apiFileLocation && !apiFileLocation[0]) {
-  throw new Error("Missing positional argument for openapi file");
-}
+const log = require("simple-node-logger").createSimpleLogger();
 
-apiFileLocation = apiFileLocation[0];
+// Override default logger
+global.console = log;
+log.setLevel("warn");
 
-let doc;
-if (apiFileLocation) {
-  doc = parse(readFileSync(apiFileLocation, "utf8"));
-} else {
-  throw new Error("Missing location. Please set OPENAPI_FILE env variable");
-}
+let { doc, apiFileLocation } = getAPI()();
 
 doc = applyOneOfTransformations(doc);
 doc = applyAllOfTransformations(doc);
 
 doc = applyModelNameTransformations(doc, "ApiAtlas", "View");
 
-writeFileSync(apiFileLocation, stringify(doc));
+saveAPI(doc, apiFileLocation);
